pyproject.toml
[tool.poetry]
name = "chatgpt-api-wrapper"
version = "0.1.0"
description = "Better prompt engineering around OpenAI API"
authors = ["avrtt <avrtt@tuta.io>"]
license = "MIT"
readme = "README.md"
packages = [{include = "chatgpt_api_wrapper", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "2.0"
python-dotenv = "^1.0.0"
openai = "^0.27.8"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
